probably nonfunctional background blur

+++ src/TweakOptions.h
@@ -3,7 +3,7 @@
 @interface TweakOptions: NSObject
 @property (nonatomic) BOOL VapourEnabled;
 @property (nonatomic) CGFloat VapourOpacity;
-@property (nonatomic) CGFloat VapourBlur;
+@property (nonatomic) uint32_t VapourBlur;
 @property (nonatomic) BOOL VapourOverrideColours;
 @property (nonatomic) BOOL VapourDisableBorder;
 @property (nonatomic) BOOL VapourDebug;
